% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqMI.semtest.R
\name{printInvarianceResult}
\alias{printInvarianceResult}
\title{Internal function to format the results of measurement invariance tests}
\usage{
printInvarianceResult(FIT, fit.measures, method, quiet = FALSE)
}
\arguments{
\item{FIT}{parameter needed by printInvarianceResult().}

\item{fit.measures}{parameter needed by printInvarianceResult().}

\item{method}{parameter needed by printInvarianceResult(). Default at "satorra.bentler.2001".}

\item{quiet}{If \code{quiet=FALSE} (default), a summary is printed out containing an overview of the different models that are fitted, together with some model comparison tests and fit measures. The results of equivalence testing will also be printed if equivalence testing is used. If \code{quiet=TRUE}, no summary is printed but results will be stored in the object.}
}
\description{
Internal function to format the results of measurement invariance tests
}
