% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{printInvarianceResult}
\alias{printInvarianceResult}
\title{Internal function to format the results of measurement invariance tests}
\usage{
printInvarianceResult(FIT, fit.measures, method, quiet = FALSE)
}
\arguments{
\item{FIT}{a list of fitted lavaan models.}

\item{fit.measures}{a vector of names of fit indices, passed to \code{lavaan:fitMeasures}.}

\item{method}{a vector of method names for calculating robust test statistic(s). See \code{lavaan:lavTestLRT}.}

\item{quiet}{If \code{quiet=FALSE} (default), a summary is printed out containing an overview of the different models that are fitted, together with some model comparison tests and fit measures. The results of equivalence testing will also be printed if equivalence testing is used. If \code{quiet=TRUE}, no summary is printed but results will be stored in the object.}
}
\description{
Internal function to format the results of measurement invariance tests
}
\author{
The maintainer, Ge Jiang, adapted the original source code of printInvarianceResult() in the \pkg{lavaan} and \pkg{semTools} packages written by Yves Rosseel, Sunthud Pornprasertmanit, and Terrence D. Jorgensen (permission obtained).
}
